Factor out `ButtonLike`
