+++ crates/ui2/src/components/button/button.rs
@@ -1,0 +1,4 @@
+#[derive(IntoElement)]
+pub struct Button5 {
+    base: ButtonLike,
+}
